+++ Projects/JSONAdminApi/Controllers/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData.Controllers.js
@@ -1,15 +1,17 @@
-let Repos = require("../../../../../../../../../Repository/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemAndData");
+let Repos = require("../../../../../../../../../Repository/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData");
 
 let PostFunc = async (req, res) => {
    let LocalDataPk = req.KeshavSoft.DataPk;
    let LocalFolderName = req.body.FolderName;
    let LocalFileName = req.body.FileName;
+   let LocalItemName = req.body.ItemName;
    let LocalData = req.body.NewData;
 
    let LocalFromRepo = await Repos.PostFunc({
       DataPK: LocalDataPk,
       inFolderName: LocalFolderName,
       inFileName: LocalFileName,
+      inItemName: LocalItemName,
       inNewData: LocalData
    });
