+++ Projects/JSONAdminApi/Dal/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData.js
@@ -1,11 +1,12 @@
-let CommonDatasupply = require("../../../../../../../../../../../DataSupply/Fs/Config/JSONFolder/DataPkAsFolder/DataFolder/UserFolder/UserJsonFile/ItemName/PushData/ItemNameAndData");
+let CommonDatasupply = require("../../../../../../../../../../../DataSupply/Fs/Config/JSONFolder/DataPkAsFolder/DataFolder/UserFolder/UserJsonFile/ItemName/PushData/ItemNameWithData");
 //let CommonDatasupplyGetFunc = require("../../../../../../../../../../../DataSupply/Fs/Config/JSONFolder/DataPkAsFolder/DataFolder/getDirectoriesWithCheckAndDelete");
 
 let path = require("path");
 
-let PostFunc = async ({ DataPK, inFolderName, inFileName,inNewData }) => {
+let PostFunc = async ({ DataPK, inFolderName, inFileName, inItemName, inNewData }) => {
     let LocalFolderName = inFolderName;
     let LocalFileName = inFileName;
+    let LocalItemName = inItemName;
     let LocalData = inNewData;
 
     if (DataPK > 0) {
@@ -13,6 +14,7 @@ let PostFunc = async ({ DataPK, inFolderName, inFileName,inNewData }) => {
             DataPK: DataPK,
             FolderName: LocalFolderName,
             FileName: path.parse(LocalFileName).name,
+            ItemName: LocalItemName,
             NewData: LocalData
         });
     };
