+++ Projects/JSONAdminApi/Repository/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData.js
@@ -1,13 +1,15 @@
-let Dal = require("../../../../../../../../../Dal/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemAndData");
+let Dal = require("../../../../../../../../../Dal/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData");
 
 exports.PostFunc = async ({ DataPK, inFolderName, inFileName, inItemName, inNewData }) => {
     let LocalFolderName = inFolderName;
     let LocalFileName = inFileName;
+    let LocalItemName = inItemName;
     let LocalData = inNewData;
 
     return await Dal.PostFunc({
         DataPK, inFolderName: LocalFolderName,
         inFileName: LocalFileName,
+        inItemName: LocalItemName,
         inNewData: LocalData
     });
 };
