--- Projects/JSONAdminApi/Routes/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemAndData.js
@@ -1,12 +1,0 @@
-var express = require('express');
-var router = express.Router();
-let CommonControllers = require("../../../../../../../../../Controllers/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemAndData.Controllers");
-let CommonMiddlewares = require("../../../../../../../../../Middlewares/AdminApi/AsTree/Json/UserFolders/DataFolder/FileInFolder/ItemName/CreateNew/CreateItemAndData");
-
-router.post('/', CommonMiddlewares.PostFunc, CommonControllers.PostFunc);
-router.delete('/', CommonMiddlewares.DeleteFunc, CommonControllers.DeleteFunc);
-router.get('/', CommonControllers.GetFunc);
-
-module.exports = router;
-
-
