+++ Projects/JSONAdminApi/Routes/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData.js
@@ -1,0 +1,12 @@
+var express = require('express');
+var router = express.Router();
+let CommonControllers = require("../../../../../../../../../Controllers/AdminApi/AsTree/Json/UserFolders/DataFolder/FileinFolder/ItemName/CreateNew/CreateItemWithData.Controllers");
+let CommonMiddlewares = require("../../../../../../../../../Middlewares/AdminApi/AsTree/Json/UserFolders/DataFolder/FileInFolder/ItemName/CreateNew/CreateItemWithData");
+
+router.post('/', CommonMiddlewares.PostFunc, CommonControllers.PostFunc);
+router.delete('/', CommonMiddlewares.DeleteFunc, CommonControllers.DeleteFunc);
+router.get('/', CommonControllers.GetFunc);
+
+module.exports = router;
+
+
